corrected return types for getXspacing() type functions.

+++ src/AdLem3D.cxx
@@ -284,21 +284,21 @@ int AdLem3D::getZnum() const
 
 #undef __FUNCT__
 #define __FUNCT__ "getXspacing"
-int AdLem3D::getXspacing() const
+double AdLem3D::getXspacing() const
 {
     return mBrainMask->GetSpacing()[0];
 }
 
 #undef __FUNCT__
 #define __FUNCT__ "getYspacing"
-int AdLem3D::getYspacing() const
+double AdLem3D::getYspacing() const
 {
     return mBrainMask->GetSpacing()[1];
 }
 
 #undef __FUNCT__
 #define __FUNCT__ "getZspacing"
-int AdLem3D::getZspacing() const
+double AdLem3D::getZspacing() const
 {
     return mBrainMask->GetSpacing()[2];
 }
